+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cli: add tables import command
 - web: add csv export to query views
 - web: add table import form
+- allow environment variable substitution in YAML configuration
 
 ### Changed
 - cli: table uri lazy validation in app configuration